requirements.txt
pyserial
matplotlib
MinimalModbus